tried to resolve delayed logout bug
Turns out there was a fix for this: running bot.run() instead of
bot.start() makes the bot log out instantly in KeyboardInterrupt.
Using .run() takes care of the event handler automatically - appearently
when discord handles this themselves logout is instant. But we dont want
that since we are going to be running alot of other asynchronous
operations in alter stages.
This is appearently a well documented bug with discords websockets not
working properly with start(). I think we're gonna have to live with
this one..