LLM changes
- Hardcoded context window for Mistral-Nemo to 100k.
Appearently this model has a bug where it reports 1M in max context
window. This gives the error message stating that I dont have 160GB of
VRAM. Which is technically correct..

- Changed system prompt to mistral version.
I feel like the sys_prompt dir is going to see alot of commits in the future.

- Increased model persistance.
Model was only staying in memory for 5 minutes. This has been increased
to 30. We can set this to negative and have it be in memory indefinately
when we have implemented the cleanup process.

- Added a function to remove model from memory.
Since we want indefinate persistent memory we should include this
function in the cleanup process.